feat(ChatComposerDrawer): rename FollowUpQuestions template to Feedback
- Rename ChatComposerDrawerFollowUpQuestions template to
  ChatComposerDrawerFeedback with updated URL slug
- Simplify to single feedback prompt: "Do you want to proceed?"
  with three selectable options (A/B/C)
- Collapsed label: "User feedback requested"
- Remove navigation arrows, "Skip all" footer action, custom placeholder
- Update storybook story to match new Feedback example
