+++ packages/cli/templates/blocks/components/ChatComposerDrawer/ChatComposerDrawerFeedback.doc.mjs
@@ -1,0 +1,9 @@
+/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
+export const doc = {
+  type: 'block',
+  name: 'ChatComposerDrawer — Feedback',
+  description: 'Chat composer drawer with a feedback prompt and selectable lettered options. Use for user confirmation workflows that require explicit action before proceeding.',
+  isReady: true,
+  aspectRatio: 4 / 3,
+  componentsUsed: ['ChatComposer', 'List', 'Badge', 'Text', 'Layout'],
+};
