--- packages/cli/templates/blocks/components/ChatComposerDrawer/ChatComposerDrawerFollowUpQuestions.doc.mjs
@@ -1,9 +1,0 @@
-/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'block',
-  name: 'ChatComposerDrawer — Follow-Up Questions',
-  description: 'Chat composer drawer with navigable follow-up questions and selectable lettered options. Use for multi-step clarification workflows.',
-  isReady: true,
-  aspectRatio: 4 / 3,
-  componentsUsed: ['ChatComposer', 'Button', 'Icon', 'List', 'Badge', 'Text', 'Layout'],
-};
